@@ -5,13 +5,12 @@
     "type": "cakephp-plugin",
     "license": "MIT",
     "require": {
-        "php": "^7.2|^8.0",
-        "phpoffice/phpspreadsheet": "^1.1",
-        "cakephp/filesystem": "~4.0",
-        "cakephp/orm": "~4.0"
+        "php": ">=8.1",
+        "phpoffice/phpspreadsheet": "^2.0",
+        "cakephp/orm": "~5.0"
     },
     "require-dev": {
-        "cakephp/cakephp": "~4.0",
+        "cakephp/cakephp": "~5.0",
         "cakephp/cakephp-codesniffer": "*",
         "phpunit/phpunit": "^9.5",
         "phpstan/phpstan": "^1.9"
